+++ neopdf/tests/pdf.rs
@@ -301,3 +301,40 @@ fn test_pdf_download() {
     assert!(gluon_as.is_finite());
     assert!(gluon_xf.is_finite());
 }
+
+#[test]
+pub fn test_xfxq2_cheby_batch() {
+    let pdf = PDF::load("MAP22_grids_FF_Km_N3LL.neopdf.lz4", 0);
+
+    let nb_points: usize = 20;
+    let generate_points = |min: f64, max: f64| -> Vec<f64> {
+        (0..nb_points)
+            .map(|i| min + i as f64 * (max - min) / (nb_points as f64 - 1.0))
+            .collect()
+    };
+    let kts: Vec<f64> = generate_points(1e-4, 5.0);
+    let xs: Vec<f64> = generate_points(1e-1, 1.0);
+    let q2s: Vec<f64> = generate_points(1.0, 1e4);
+
+    let mut flatten_points: Vec<Vec<f64>> = Vec::new();
+    for &kt in &kts {
+        for &x in &xs {
+            for &q2 in &q2s {
+                flatten_points.push(vec![kt, x, q2]);
+            }
+        }
+    }
+
+    let points_interp: Vec<&[f64]> = flatten_points.iter().map(Vec::as_slice).collect();
+    let slice_points: &[&[f64]] = &points_interp;
+
+    let ids: Vec<i32> = (-3..=3).filter(|&x| x != 0).collect();
+    for &pid in &ids {
+        let results_batch = pdf.xfxq2_cheby_batch(pid, slice_points);
+        let results_seq: Vec<f64> = slice_points.iter().map(|p| pdf.xfxq2(pid, p)).collect();
+
+        for (res_b, res_s) in results_batch.iter().zip(results_seq.iter()) {
+            assert!((res_b - res_s).abs() < LOW_PRECISION);
+        }
+    }
+}