+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Added a logic to compute Chebyshev interpolations in batches (https://github.com/Radonirinaunimi/neopdf/pull/64).
 - Added proper LHAPDF drop-in compatibility layer for no-code migration.
 - Added an interface to the Wolfram Language to allow Rust APIs to be called in
   Mathematica.